CGContextSetLineWidth defines the line width as a CGFloat

+++ src/CKSparkline.h
@@ -7,14 +7,14 @@
 	UIColor *highlightedLineColor;
 	NSArray *data;
 	NSArray *computedData;
-	int lineWidth;
+	CGFloat lineWidth;
 }
 
 
 @property (readonly) BOOL selected;
 @property (nonatomic, retain) UIColor *lineColor;
 @property (nonatomic, retain) UIColor *highlightedLineColor;
-@property (nonatomic) int lineWidth;
+@property (nonatomic) CGFloat lineWidth;
 @property (readonly) NSArray *data;
 @property (readonly) NSArray *computedData;
